feature:페이지 코드 컨벤션 수정

+++ src/main/java/com/kh/wellness/admin/place/model/dao/AdminPlaceMapper.java
@@ -12,5 +12,5 @@ public interface AdminPlaceMapper {
 
 	long countPlaces();
 
-	List<AdminPlaceListResponse> selectPlaces(@Param("offset") int offset, @Param("size") int size);
+	List<AdminPlaceListResponse> selectPlaces(@Param("offset") Long offset, @Param("size") int size);
 }
